Updated with code dealing with table view and collection view, expanded tip generator to contain future topics

+++ iOSMobileTips/TipsGeneratorArray.h
@@ -15,6 +15,6 @@
 -(NSString*)getTitleAtIndex:(NSUInteger)index;
 -(NSString*)getTipDescriptionAtIndex:(NSUInteger)index;
 -(NSString*)getCodeSegmentAtIndex:(NSUInteger)index;
--(int)getCount;
+-(NSUInteger)getCount;
 
 @end
